Set the keybinding globally rather than in a minor mode.
This allows org-mode's `org-cycle' command to fall-back to `smart-tab' if it
decides not to do any "org-mode-y" functions. Rather than have `smart-tab' try
to deal with what it should do if it is called within a the minibuffer, prevent
`smart-tab-mode' from starting in a minibuffer.

This is a better way of organizing things, since if a major or minor mode wants
to override the function of tab, then it should have priority over
`smart-tab'. If that mode wants to provide a fallback (like org-mode does if it
cannot cycle at the current point), it can use `global-key-binding' to call
`smart-tab'.

Signed-off-by: Dan Hackney <chrono325@gmail.com>